// src/generated/models/staff-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records - Appointment scheduling and management - Triage, vitals, and examination - Prescription and investigation management - IPD/OPD workflow 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StaffStatus } from './staff-status';

export interface StaffResponse {
    'id'?: number;
    'emailId'?: string;
    'fullName'?: string;
    'contactNumber'?: string;
    'status'?: StaffStatus;
    'roles'?: Array<string>;
    'createdAt'?: string;
    'updatedAt'?: string;
}